feat: adiciona tratamento de exceções

+++ src/storage/group/groupCreate.ts
@@ -1,11 +1,19 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import { GROUP_COLLECTION } from '@storage/storageConfig';
+import { AppError } from '@utils/AppError';
 import { groupGetAll } from './groupsGetAll';
+import { GROUP_COLLECTION } from '@storage/storageConfig';
 
 export async function groupCreate(newGroup: string) {
   try {
     const storageGroups = await groupGetAll();
+    const groupAlreadyExists = storageGroups.includes(newGroup);
+
+    if (groupAlreadyExists) {
+      throw new AppError('Já existe uma turma com esse nome');
+    }
+
     const storage = JSON.stringify([...storageGroups, newGroup]);
+
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
   } catch (error) {
     throw error;
